requirements.txt

bson==0.5.8
camelot-py==0.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudinary==1.20.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.1.7
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangotoolbox==1.8.0
djongo==1.3.1
djoser==2.0.3

et-xmlfile==1.0.1
idna==2.9
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
openapi-codec==1.3.2
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.3
pdfminer.six==20200517
Pillow==7.1.2
pycryptodome==3.9.7
PyJWT==1.7.1
pymongo==3.10.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0.1
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.9
